requirements.txt
gradio==5.8.0
huggingface_hub==0.16.4
langchain==0.0.27
langchain_community==0.3.11
langchain_huggingface==0.1.2
python-dotenv==1.0.1
torch==1.13.1
transformers==4.30.2
